python/pyproject.toml
[project]
name = "mlops"
version = "0.1.0"
description = "MLOps Python layer for ml-offload-api — optimization, tracking, LLM client, reasoning"
requires-python = ">=3.13"
dependencies = [
  "mlflow>=2.14.0",
  "temporalio>=1.6.0",
  "dspy-ai>=2.4.0",
  "anthropic>=0.40.0",
  "openai>=1.54.0",
  "tenacity>=8.2.3",
  "pyyaml>=6.0.3",
  "grpcio>=1.62.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mlops"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
